:pencil2: 위/경도 및 가게평점 데이터 타입을 Float에서 Double로 변경

+++ src/main/java/kr/bb/store/domain/store/controller/StoreController.java
@@ -53,7 +53,7 @@ public class StoreController {
     }
 
     @GetMapping("/map/location?lat={lat}&lon={lon}")
-    public ResponseEntity getNearbyStores(@PathVariable Float lat, @PathVariable Float lon) {
+    public ResponseEntity getNearbyStores(@PathVariable Double lat, @PathVariable Double lon) {
         return ResponseEntity.ok().body(storeService.getNearbyStores(lat,lon));
     }
 }
